==> 0_download_qwen.ipynb <==
from transformers import AutoProcessor, AutoModelForCausalLM # The `local_files_only=False` ensures it downloads if missing.
from huggingface_hub import snapshot_download

# Qwen3-4b
This is the base qwen3-4b modell. This is the pretrained model we finetune for safety allignment. 

#-----------------------Base Qwen-----------------------
# processor = AutoProcessor.from_pretrained(
#     "Qwen/Qwen3-4B",
#     cache_dir="./models/qwen3"
    
# )
# model = AutoModelForCausalLM.from_pretrained(
#     "Qwen/Qwen3-4B",
#     cache_dir="./models/qwen3"
# )

# snapshot_download(
#     repo_id="Qwen/Qwen3-4B",
#     local_dir="./models/qwen3/Qwen3-4B",
# )

processor = AutoProcessor.from_pretrained("./models/qwen3/Qwen3-4B", local_files_only=True)
model = AutoModelForCausalLM.from_pretrained("./models/qwen3/Qwen3-4B", local_files_only=True)

messages = [
    {"role": "user", "content": "Who are you?"},
]
inputs = tokenizer.apply_chat_template(
	messages,
	add_generation_prompt=True,
	tokenize=True,
	return_dict=True,
	return_tensors="pt",
).to(model.device)

outputs = model.generate(**inputs, max_new_tokens=40)
print(tokenizer.decode(outputs[0][inputs["input_ids"].shape[-1]:]))

# Qwen3-4b-SafeRL
This is a pretrained and safety aligned version of the qwen3-4b modell. It will serve as a comparison to our fine-tuned model. 

snapshot_download(
     repo_id="Qwen/Qwen3-4B-SafeRL",
     local_dir="./models/qwen3/Qwen3-4B-SafeRL",
)
